Cut more wobbling code down
JAVA-4751
